trx/output: extract lights.c and state.c
