+++ src/libtrx/include/libtrx/game/output.h
@@ -4,6 +4,7 @@
 #include "./output/common.h"
 #include "./output/const.h"
 #include "./output/draw.h"
+#include "./output/lights.h"
 #include "./output/state.h"
 #include "./output/textures.h"
 #include "./output/types.h"
